requirements.txt
langchain>=0.1.0
langchain-anthropic>=0.1.1
python-dotenv>=1.0.0
notebook>=7.0.0 